backend/requirements.txt
fastapi
uvicorn
boto3
pydantic
reportlab
python-dotenv
